properties/requirements.txt
django==4.0.4
djangorestframework==3.13.1
drf-yasg==1.20.0
requests==2.27.1
django-extensions==3.1.5
